refactor integration tests
